test: cover incident correlation failure paths
