+++ .groom/review-scores.ndjson
@@ -1,0 +1,1 @@
+{"date":"2026-03-31","pr":null,"correctness":8,"depth":8,"simplicity":8,"craft":8,"verdict":"ship"}
